backend/requirements.txt
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
appdirs==1.4.4
argcomplete==3.6.2
bcrypt==4.0.1
bidict==0.23.1
cffi==2.0.0
click==8.1.8
cryptography==46.0.2
dnspython==2.7.0
ecdsa==0.19.1
escpos==1.9
eventlet==0.33.3
exceptiongroup==1.3.0
fastapi==0.104.1
future==1.0.0
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
Pillow==10.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.23
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pyserial==3.5
python-barcode==0.16.1
python-dotenv==1.0.0
python-engineio==4.12.3
python-jose==3.3.0
python-multipart==0.0.6
python-socketio==5.10.0
pyusb==1.3.1
PyYAML==6.0.3
qrcode==8.2
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.15.0
uvicorn==0.24.0
uvloop==0.21.0
viivakoodi==0.8.0
watchfiles==1.1.0
websockets==15.0.1
wsproto==1.2.0
